api/views.py
from rest_framework import viewsets
from accounts.models import Customer
from api.serializers import CustomerSerializer


class CustomerViewSet(viewsets.ModelViewSet):
    queryset = Customer.objects.all()
    serializer_class = CustomerSerializer
