requirements.txt
django==1.8.2
django-localflavor==1.1
pillow==2.8.1
djangorestframework==3.1.2